requirements.txt
openai
streamlit
beautifulsoup4
requests
faiss-cpu
nltk
transformers
torch
